+++ tests/test_get_phpunit_executable.py
@@ -1,7 +1,41 @@
 import os
 
+from PHPUnitKit.plugin import PHPUnit
 from PHPUnitKit.plugin import _get_phpunit_executable
 from PHPUnitKit.tests import unittest
+
+
+class TestPHPUnitGetPHPUnitExecutable(unittest.ViewTestCase):
+
+    def test_get_user_phpunit_executable(self):
+        phpunit = PHPUnit(self.view.window())
+        self.view.settings().set('phpunit.executable', 'fizz')
+        self.assertEqual('fizz', phpunit.get_phpunit_executable(working_dir='foo'))
+
+    def test_get_user_phpunit_executable_is_filtered(self):
+        phpunit = PHPUnit(self.view.window())
+        home = os.path.expanduser('~')
+        self.view.settings().set('phpunit.executable', '~')
+        self.assertEqual(home, phpunit.get_phpunit_executable(working_dir='foo'))
+        self.view.settings().set('phpunit.executable', '$HOME')
+        self.assertEqual(home, phpunit.get_phpunit_executable(working_dir='foo'))
+
+    def test_get_user_phpunit_executable_allows_executable_as_list(self):
+        phpunit = PHPUnit(self.view.window())
+        self.view.settings().set('phpunit.executable', ['fizz', 'buzz'])
+        self.assertEqual(['fizz', 'buzz'], phpunit.get_phpunit_executable(working_dir='foo'))
+
+    def test_get_user_phpunit_executable_as_list_is_filtered(self):
+        phpunit = PHPUnit(self.view.window())
+        home = os.path.expanduser('~')
+        self.view.settings().set('phpunit.executable', ['~'])
+        self.assertEqual([home], phpunit.get_phpunit_executable(working_dir='foo'))
+        self.view.settings().set('phpunit.executable', ['~', '~'])
+        self.assertEqual([home, home], phpunit.get_phpunit_executable(working_dir='foo'))  # noqa: E501
+        self.view.settings().set('phpunit.executable', ['$HOME'])
+        self.assertEqual([home], phpunit.get_phpunit_executable(working_dir='foo'))
+        self.view.settings().set('phpunit.executable', ['$HOME', '$HOME'])
+        self.assertEqual([home, home], phpunit.get_phpunit_executable(working_dir='foo'))  # noqa: E501
 
 
 class TestGetPHPUnitExecutable(unittest.TestCase):